requirements.txt
bcrypt==5.0.0
blinker==1.8.2
click==8.1.8
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==5.0.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
PyJWT==2.9.0
pymongo==4.10.1
python-decouple==3.8
python-dotenv==1.0.1
Werkzeug==3.0.6
zipp==3.20.2